+++ tests/test_utility.py
@@ -8,11 +8,14 @@ from tests.conftest import graph_equal
 import networkx as nx
 from functools import wraps
 
+
 @pytest.fixture
 def func():
     def example_func(a, c, b=2, *args, d, e=10, **kwargs):
         return
+
     return example_func
+
 
 @pytest.mark.parametrize(
     "parameter, result",
@@ -128,34 +131,34 @@ def test_param_counter_add_returns(mmodel_G):
     assert counter == {"a": 1, "b": 2, "c": 4, "d": 1, "e": 1, "f": 1, "g": 2}
 
 
-# def test_modify_subgraph_terminal(mmodel_G):
-#     """Test redirect edges based on subgraph and subgraph node
+def test_modify_subgraph_terminal(mmodel_G):
+    """Test redirect edges based on subgraph and subgraph node
 
-#     This test specifically the terminal node
-#     """
+    This tests specifically the terminal node
+    """
 
-#     subgraph = mmodel_G.subgraph(["multiply", "poly"])
+    subgraph = mmodel_G.subgraph(["multiply", "poly"])
 
-#     def mock_obj(c, x, y):
-#         return
+    def func(c, x, y):
+        return
 
-#     graph = util.modify_subgraph(mmodel_G, subgraph, "test", mock_obj)
+    graph = util.modify_subgraph(mmodel_G, subgraph, "test", func)
 
-#     # a copy is created
-#     assert graph != mmodel_G
-#     assert "test" in graph
+    # a copy is created
+    assert graph != mmodel_G
+    assert "test" in graph
 
-#     assert graph.nodes["test"] == {
-#         "base_func": mock_obj,
-#         "modifiers": [],
-#         "func": mock_obj,
-#         "output": None,
-#         "sig": inspect.signature(mock_obj),
-#     }
+    assert graph.nodes["test"] == {
+        "base_func": func,
+        "modifiers": [],
+        "func": func,
+        "output": None,
+        "sig": inspect.signature(func),
+    }
 
-#     # Test the edge attributes
-#     assert graph.edges["add", "test"]["val"] == "c"
-#     assert graph.edges["subtract", "test"]["val"] is None
+    # Test the edge attributes
+    assert graph.edges["add", "test"]["val"] == "c"
+    assert graph.edges["subtract", "test"]["val"] == "e"
 
 
 def test_modify_subgraph_middle(mmodel_G):
@@ -190,68 +193,23 @@ def test_modify_subgraph_middle(mmodel_G):
     assert graph.edges["test", "multiply"]["val"] == "e"
 
 
-def test_modify_subgraph_none_returns(mmodel_G):
-    """Test subgraph modification when returns are not specified
-
-    This test specifically the middle node
-    """
-
-    subgraph = mmodel_G.subgraph(["poly", "multiply"])
-
-    def mock_obj(c, x, y):
-        return
-
-    mock_obj.output = "e"
-
-    graph = util.modify_subgraph(mmodel_G, subgraph, "test", mock_obj)
-
-    assert graph.nodes["test"] == {
-        "base_func": mock_obj,
-        "modifiers": [],
-        "func": mock_obj,
-        "output": "e",
-        "sig": inspect.signature(mock_obj),
-    }
-
-
-# def test_modify_subgraph_none_returns_fails(mmodel_G):
-#     """Test subgraph modification when returns are not specified
-
-#     The method fails when the subgraph node does not have a
-#     return attribute
-#     """
-
-#     subgraph = mmodel_G.subgraph(["poly", "multiply"])
-
-#     def mock_obj(c, x, y):
-#         return
-
-#     with pytest.raises(Exception, match="'returns' not defined"):
-#         util.modify_subgraph(mmodel_G, subgraph, "test", mock_obj)
-
-
 def test_modify_node(mmodel_G):
     """Test modify_node
 
     Test if the node have the correct signature and result
     """
 
-    def mock_obj(c, x, y):
-        return x + y
-
-    def mod(func):
+    def mod(func, a):
         @wraps(func)
         def wrapped(*args, **kwargs):
-            return func(*args, **kwargs) + 1
+            return func(*args, **kwargs) + a
 
         return wrapped
 
-    mod_G = util.modify_node(mmodel_G, "subtract", [(mod, {})], "g")
+    mod_G = util.modify_node(mmodel_G, "subtract", modifiers=[(mod, {'a': 1})])
 
     # add one to the final value
     assert mod_G.nodes["subtract"]["func"](1, 2) == 0
-    # make sure the edge value is updated
-    assert mod_G["subtract"]["multiply"]["val"] == "g"
 
 
 def test_is_node_attr_defined():